AP_BattMonitor: Allow temperature data to be valid over a longer time range
Allows temperature data to be considered valid if it was recieved within the
 last 5 seconds.

This is needed because the Solo implementation suffers from an incredibly lossy
I2C bus. Testing shows that only ~25% of all attempts to read data from the
battery on Solo will succeed, which created a very noisy looking teperature
reading. Because of this bus the Solo implementation no longer requests the
pack voltage, it instead works by requesting the cells (since it is a single
block transfer) and adding them up to provide the pack voltage. This saves us
one transaction per loop