AP_Notify: Added GPS Fusion flag from AHRS
The existing notify flags for GPS all report on the actual GPS having a
lock or not. But this leaves a lot to be desired. The EKF still decides
whether or not the GPS fix is usable, regardless of it being a 3D fix or
not. So this reports on the filter state of the EKF in AP_AHRS_NavEKF.
If the GPS location is actually usable and being used by the EKF, this
flag will be true.  If the GPS position is being rejected by the EKF,
this flag will be false.  The use case for this is driving more
meaningful system feedback to the operator. This flag can now drive
tones, LEDs, and GCS feedback that the operator currently does not have.