AP_Notify: Complete rework of notify device selection
Jaime did the hard work on this one.  He created NTF_BRD_TYPE and
reworked notify device selection to take place on init rather than on
compile like before.  Notify device selection is now completely dynamic
so the user can select what applies to them and not bother with what
doesn't apply. This should save some memory here and there too.  I then
modified it a bit, creating NTF_OLED_THEME.  This will allow the user to
enable/disable the OreoLED driver. And it also allows you to select
between aircraft and rover lighting themes. This allows the Solo to use
the OreoLEDs, and doesn't waste the memory on vehicles not equipped with
Oreo LEDs.

There is also a new notify flag for GPS Fusion.  This flag is true when
the EKF is happy with the GPS, actively using it for position
information.