Rework device selection, drop notify board type
Per discussion on the 5/29/2017 dev call... The need for a notify board
type is not really needed since the only thing that needs to be
dynamically switched is Oreo LEDs.  The NTF_OLED_THEME parameter handles
that entirely.  Reworked code so all other notify devices are selected
by preprocessor directives.  This maintains Jaime's notify device
structure to do it.